+++ src/Products/MeetingCommunes/tests/testCustomViews.py
@@ -8,6 +8,7 @@
 from collective.contact.plonegroup.utils import get_plone_group_id
 from DateTime import DateTime
 from imio.history.utils import getLastWFAction
+from itertools import chain
 from plone import api
 from plone.app.textfield import RichTextValue
 from plone.dexterity.utils import createContentInContainer
@@ -883,6 +884,20 @@ class testCustomViews(MeetingCommunesTestCase):
             unrestricted=True)
         self.assertEqual([item.getProposingGroup() for item in grouped_items],
                          [self.vendors_uid, self.vendors_uid, self.vendors_uid])
+        # itemUids, unrestricted will return everything unless len given itemUids
+        # is < every visible items meaining user filtered meeting in the UI
+        # when passing itemUids of every visible items, we get more
+        all_unrestricted_grouped_items = list(
+            chain.from_iterable([items for gp_title, items in unrestricted_grouped_items]))
+        self.assertEqual(len(itemUids), 3)
+        self.assertEqual(len(all_unrestricted_grouped_items), 7)
+        itemUids = [item.UID() for item in all_unrestricted_grouped_items[0:4]]
+        self.assertEqual(len(itemUids), 4)
+        unrestricted_grouped_items = helper.get_grouped_items(
+            itemUids, group_by='proposingGroup', unrestricted=True)
+        all_unrestricted_grouped_items = list(
+            chain.from_iterable([items for gp_title, items in unrestricted_grouped_items]))
+        self.assertEqual(len(all_unrestricted_grouped_items), 4)
 
     def test_get_grouped_items_additional_catalog_query(self):
         self.changeUser('pmManager')
@@ -959,12 +974,12 @@ class testCustomViews(MeetingCommunesTestCase):
         cfg = self.meetingConfig
         cfg.setUseGroupsAsCategories(True)
 
-        def create_and_validate_item(creator, preffered_meeting=None):
+        def create_and_validate_item(creator, preferred_meeting=None):
             self.changeUser(creator)
             item = self.create('MeetingItem')
             self.validateItem(item)
-            if preffered_meeting:
-                item.setPreferredMeeting(preffered_meeting)
+            if preferred_meeting:
+                item.setPreferredMeeting(preferred_meeting)
             return item
 
         def get_item_view(item):