+++ CHANGES.rst
@@ -43,6 +43,12 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
   as `MeetingItem.getCategory` does no more return the `proposingGroup` when
   `MeetingConfig.useGroupsAsCategories` is True.
   [gbastien]
+- Renamed `MCItemDocumentGenerationHelperView.print_deliberation` to
+  `MCItemDocumentGenerationHelperView.print_full_deliberation` as
+  `print_deliberation` already exists in PloneMeeting and is used by
+  `ItemDocumentGenerationHelperView.deliberation_for_restapi`.
+  Added Migrate_To_4200._fixPODTemplatesInstructions to fix it in POD templates.
+  [gbastien]
 
 4.2b9 (2021-01-26)
 ------------------